Reorganize components directory
